server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Mykyta Prokaiev <mykytaprokaiev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
uwsgi = "^2.0.23"
psycopg2 = "^2.9.9"
django = "^5.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.31.0"
django-cors-headers = "^4.3.1"
djangorestframework-api-key = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
